dividedTheBlocks
